HHH-10873 Type systems consolidation + org.hibernate.type refactoring
re-work BasicType, Column and BasicValuedNavigable as discussed

+++ hibernate-core/src/main/java/org/hibernate/cfg/IndexColumn.java
@@ -23,7 +23,7 @@ public class IndexColumn extends Ejb3Column {
 	public IndexColumn(
 			boolean isImplicit,
 			String sqlType,
-			int length,
+			long length,
 			int precision,
 			int scale,
 			String name,
